Using cstdint for most integer types now. Added level renderer

+++ src/shared/config.hpp
@@ -1,9 +1,11 @@
 #pragma once
 
-namespace Config 
+#include <cstdint>
+
+namespace Config
 {
     static const float MILLIS_PER_TICK = 33.333333f;
 
-    static const unsigned int MAX_PACKET_SIZE = 2048;
-    static const unsigned int DEFAULT_PORT = 12345;
+    static const uint32_t MAX_PACKET_SIZE = 2048;
+    static const uint16_t DEFAULT_PORT = 12345;
 }
